Split tests into their own files
